requirements.txt
flask
python-dotenv
requests
supabase
openai>=1.0.0
pytz